Move event_binner out of cell_group.hpp (#229)
* Move `event_binner` class to own header file `event_binner.hpp` with implementation in `event_binner.cpp`.
* Move `event_binner` unit tests to own source file `test_event_binner.cpp`.
