+++ src/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(BASE_SOURCES
     common_types_io.cpp
     cell.cpp
+    event_binner.cpp
     morphology.cpp
     parameter_list.cpp
     profiling/memory_meter.cpp
